Small fix, moving worldgen.js to /lib
