+++ lib/serverPlugins/world.js
@@ -1,6 +1,6 @@
 var Chunk = require('prismarine-chunk')(require("../version"));
 var World = require('prismarine-world');
-var WorldGen = require('./worldgen');
+var WorldGen = require('../worldgen');
 var Vec3 = require('vec3');
 
 module.exports=inject;
@@ -39,7 +39,6 @@ function inject(serv) {
           }
         }
       }
-      serv.world.setColumn(chunkX,chunkZ,chunk);
     }
   }
 }
