+++ lib/worldgen.js
@@ -1,4 +1,4 @@
-var Chunk = require('prismarine-chunk')(require("../version"));
+var Chunk = require('prismarine-chunk')(require("./version"));
 var Vec3 = require('vec3');
 var gen = require('random-seed');
